requirements.txt
ply>=3.0
python-dateutil
numpy
rpy2